refactor(shared/chat_protocol): session 서브패키지로 재구성 (3 파일 분리)
session_runtime.py + _chat_event_buffer.py 를 `session/` 서브패키지로
정리. 각 파일이 한 클래스 책임.

새 구조:
  chat_protocol/session/
  ├── __init__.py            # SessionRuntime / SessionRegistry / 상수만 export
  ├── chat_event_buffer.py   # _ChatEventBuffer (internal, init 에서 export X)
  ├── runtime.py             # SessionRuntime + DEFAULT_MAX_BACKLOG_MESSAGES
  └── registry.py            # SessionRegistry + TTL sweeper + 상수

ChatEventBuffer 는 SessionRuntime 의 implementation detail — `session/__init__`
에서 의도적으로 export 하지 않음. 직접 접근 (단위 테스트 등) 은 explicit
import (`from dev_team_shared.chat_protocol.session.chat_event_buffer import
_ChatEventBuffer`).

외부 API 호환 유지:
- `from dev_team_shared.chat_protocol import SessionRuntime, SessionRegistry` 동일
- Primary chat_handler / 테스트 코드 영향 없음 (top-level import 만 사용)
